requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.1
fastapi-jwt-auth==0.5.0
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
setuptools==68.2.2
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wheel==0.41.2
